Abstract the rateLimitedFetch logic to its own file for reusability
